+++ ts/src/config.ts
@@ -97,13 +97,17 @@ function assertEnvVarType(name: string, value: any, expectedType: EnvVarType): a
     }
 }
 function getDefaultFeeRecipient(): string {
-    const metadata = JSON.parse(fs.readFileSync(metadataPath).toString());
-    const existingDefault: string = metadata.DEFAULT_FEE_RECIPIENT;
-    const newDefault: string = existingDefault || `0xabcabc${crypto.randomBytes(17).toString('hex')}`;
-    if (_.isEmpty(existingDefault)) {
-        const metadataCopy = JSON.parse(JSON.stringify(metadata));
-        metadataCopy.DEFAULT_FEE_RECIPIENT = newDefault;
-        fs.writeFileSync(metadataPath, JSON.stringify(metadataCopy));
+    let newDefault: string = `0xabcabc${crypto.randomBytes(17).toString('hex')}`
+    if (fs.existsSync(metadataPath)) {
+      const metadata = JSON.parse(fs.readFileSync(metadataPath).toString());
+      const existingDefault: string = metadata.DEFAULT_FEE_RECIPIENT;
+      newDefault = existingDefault || newDefault;
+      if (_.isEmpty(existingDefault)) {
+          const metadataCopy = JSON.parse(JSON.stringify(metadata));
+          metadataCopy.DEFAULT_FEE_RECIPIENT = newDefault;
+          fs.writeFileSync(metadataPath, JSON.stringify(metadataCopy));
+      }
     }
+
     return newDefault;
 }
